Refactor MMTk-specific changes
. Change build mechanism to better match other Julia optional libraries.
. Reduce `#ifdef MMTK`s.
. Add `gc-common.c` to hold GC code that is common to Julia's GC and
  MMTk.
. Add `mmtk-gc.c` for MMTk-specific code.
. Add `gc-interface.h` to document the GC interface.
