+++ src/llvm-pass-helpers.cpp
@@ -226,13 +226,8 @@ namespace jl_intrinsics {
 }
 
 namespace jl_well_known {
-#ifndef MMTKHEAP
     static const char *GC_BIG_ALLOC_NAME = XSTR(jl_gc_big_alloc);
     static const char *GC_POOL_ALLOC_NAME = XSTR(jl_gc_pool_alloc);
-#else
-    static const char *GC_BIG_ALLOC_NAME = XSTR(jl_mmtk_gc_alloc_big);
-    static const char *GC_POOL_ALLOC_NAME = XSTR(jl_mmtk_gc_alloc_default_llvm);
-#endif
     static const char *GC_QUEUE_ROOT_NAME = XSTR(jl_gc_queue_root);
 
     using jl_intrinsics::addGCAllocAttributes;
